sysdump: ignore some files in /sys/kernel/tracing to avoid hang
Identified the following files in /sys/kernel/tracing which cause hang:

* /sys/kernel/tracing/trace_pipe
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/trace_pipe
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/trace_pipe_raw
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/snapshot_raw

Instead of ignoring the files explicitly like this:

  if ($file =~ "/sys/kernel/tracing/per_cpu/cpu[0-9]+/trace_pipe*") {
  if ($file =~ "/sys/kernel/tracing/trace_pipe") {

and to make sysdump more robust for future blocking files,
it seemed more obvious to ignore all files which contain '*_pipe*'.

Also moved print $file to the beginning of the block
to make it easier to debug, in case other files cause blocking.

Closes: grml/grml-hwinfo#3