sysdump: rework ignore list code and ignore further files
When getting towards /sys/kernel/tracing, the script is known to hang
at least at the following files:

* /sys/kernel/tracing/trace_pipe
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/trace_pipe
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/trace_pipe_raw
* /sys/kernel/tracing/per_cpu/cpu[0-9]+/snapshot_raw

So we need to extend the list of files to be ignored.

While at it, rework the checks and ignore a bunch of further files, that
can't be read, e.g. due to "Permission denied".

FTR, we sadly can't ignore "/sys/.*/uevent" files globally, as
there are a bunch of files that contain and also provide actual data.

Thanks: Darshaka Patirana for tracking down the /sys/kernel/tracing/* issue
Closes: grml/grml-hwinfo#3