Iterate over present disks to collect "sfdisk -d" information
AFAICT, up and until Debian/jessie the sfdisk executable could be
invoked without the need to specify any specific device name on the
command line. In "newer" versions of sfdisk this is no longer supported,
so we need to iterate over the list of present disk devices.

With this change, the sfdisk_... files can also be used standalone,
which is useful e.g. for restoring a partition table.

Development time sponsored by Sipwise GmbH