grml-hwinfo: Warn the user if grml-hwinfo is running without X server
Not all data is collected when grml-hwinfo is run without an X server.
The user is now warned.

Also improved the check if the X server is not running.

So far, we relied on the existence of the $DISPLAY variable to detect if
the X server is running.

If the variable is set but the X server is not running, we use xset
quietly to see if we can connect to the X server.

Inspired by: https://stackoverflow.com/a/11965765/2142030